+++ packages/plugins/src/index.ts
@@ -1,5 +1,3 @@
 export { default as RematchWebpackPlugin } from './rematch';
 
-export { default as WidgetsWebpackPlugin } from './widgets';
-
-export { default as RoutesWebpackPlugin } from "./routes"
+export { default as RoutesWebpackPlugin } from './routes';